Put the classes where they belong.
