+++ release-controller/commit_annotator.py
@@ -9,6 +9,7 @@ import re
 import threading
 import time
 import typing
+import urllib
 from prometheus_client import start_http_server, Gauge
 
 sys.path.append(os.path.join(os.path.dirname(__file__)))
@@ -255,6 +256,127 @@ def annotate_commits_of_branch(
         logger.info("Successfully annotated %s commits", len(commits))
     else:
         COMMITS_BEHIND.labels(branch).set(0)
+
+
+class NotReady(Exception):
+    """Exception raised when a commit is not yet annotated."""
+
+    pass
+
+
+class LocalCommitChangeDeterminator(object):
+    """Retrieves annotations from a local Git repository."""
+
+    def __init__(self, ic_repo: GitRepo):
+        """
+        Creates a new commit change determinator.
+
+        Upon creation, the freshest notes are fetched.
+        """
+        self.annotator = GitRepoAnnotator(
+            ic_repo,
+            list(CHANGED_NOTES_NAMESPACES.values()),
+            save_annotations=False,
+        )
+        self.annotator.fetch()
+
+    def commit_changes_artifact(
+        self, commit: str, os_kind: OsKind
+    ) -> CommitInclusionState:
+        """
+        Check if the os_kind (artifact) changed in the specifed commit
+        by querying the local repo for git notes populated and pushed
+        by commit annotator.
+        """
+        namespace = CHANGED_NOTES_NAMESPACES[os_kind]
+        try:
+            changed = (
+                self.annotator.get(namespace=namespace, object=commit)
+                .decode("utf-8")
+                .strip()
+            )
+        except KeyError:
+            raise NotReady(
+                f"Could not find {os_kind} label for commit {commit}. Check out commit annotator logs and runbook: https://dfinity.github.io/dre/release.html#missing-guestos-label."
+            )
+        assert changed in [
+            COMMIT_BELONGS,
+            COMMIT_DOES_NOT_BELONG,
+            COMMIT_COULD_NOT_BE_ANNOTATED,
+        ], "Expected a specific CommitInclusionState, not %r" % changed
+        return typing.cast(CommitInclusionState, changed)
+
+
+class RecreatingCommitChangeDeterminator(object):
+    """
+    Computes annotations on the fly from a local Git repository, ignoring existing annotations.
+
+    Annotations made this way are never saved or published to the Git repository.
+    """
+
+    def __init__(self, ic_repo: GitRepo):
+        """
+        Creates a new commit change determinator.
+
+        Upon creation, the freshest notes are fetched.
+        """
+        self.annotator = GitRepoAnnotator(
+            ic_repo,
+            list(CHANGED_NOTES_NAMESPACES.values()),
+            save_annotations=False,
+        )
+
+    def commit_changes_artifact(
+        self, commit: str, os_kind: OsKind
+    ) -> CommitInclusionState:
+        """
+        Check if the os_kind (artifact) changed in the specifed commit
+        by querying the local repo for git notes populated and pushed
+        by commit annotator.
+        """
+
+        _, _, belongs = compute_annotations_for_object(self.annotator, commit, os_kind)
+        assert belongs in [
+            COMMIT_BELONGS,
+            COMMIT_DOES_NOT_BELONG,
+            COMMIT_COULD_NOT_BE_ANNOTATED,
+        ], "Expected a specific CommitInclusionState, not %r" % belongs
+        return belongs
+
+
+class CommitAnnotatorClientCommitChangeDeterminator(object):
+    """Retrieves annotations from a commit annotator API server."""
+
+    def __init__(self, base_url: str):
+        self.base_url = base_url
+
+    def commit_changes_artifact(
+        self, commit: str, os_kind: OsKind
+    ) -> CommitInclusionState:
+        """
+        Check if the os_kind (artifact) changed in the specifed commit
+        by querying the commit annotator for git notes.
+        """
+        namespace = CHANGED_NOTES_NAMESPACES[os_kind]
+        url = (
+            self.base_url.rstrip("/")
+            + f"/api/v1/commit/{commit}/annotation/{namespace}"
+        )
+        try:
+            with urllib.request.urlopen(url) as response:
+                changed = response.read().decode("utf-8").strip()
+        except urllib.error.HTTPError as he:
+            if he.code == 404:
+                raise NotReady(
+                    f"Could not find {os_kind} label for commit {commit}. Check out commit annotator logs and runbook: https://dfinity.github.io/dre/release.html#missing-guestos-label."
+                ) from he
+            raise
+        assert changed in [
+            COMMIT_BELONGS,
+            COMMIT_DOES_NOT_BELONG,
+            COMMIT_COULD_NOT_BE_ANNOTATED,
+        ], "Expected a specific CommitInclusionState, not %r" % changed
+        return typing.cast(CommitInclusionState, changed)
 
 
 class APIHandler(http.server.BaseHTTPRequestHandler):
